refactor: install github via endpoint
