+++ svc/api/routes/v2_workspaces_install_github/401_test.go
@@ -6,22 +6,17 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"github.com/unkeyed/unkey/svc/api/internal/testutil"
-	handler "github.com/unkeyed/unkey/svc/api/routes/v2_apps_create_github_connection"
 )
 
-func TestCreateGithubConnectionUnauthorized(t *testing.T) {
+func TestInstallGithubUnauthorized(t *testing.T) {
 	h := testutil.NewHarness(t)
 
 	route := newRoute(h)
 	h.Register(route)
 
 	headers := http.Header{
-		"Content-Type":  {"application/json"},
 		"Authorization": {"Bearer invalid_token"},
 	}
-	res := testutil.CallRoute[handler.Request, handler.Response](h, route, headers, handler.Request{
-		Project: "payments",
-		App:     "payments-api",
-	})
+	res := callInstall(h, route, headers)
 	require.Equal(t, http.StatusUnauthorized, res.Status, "expected 401, received: %s", res.RawBody)
 }
